Warnings removed
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/core/trunk@499 77d55b08-817d-0410-979b-1fec1beb8c2e

+++ moldyn/DirectVolumeRenderer.cpp
@@ -955,7 +955,7 @@ void moldyn::DirectVolumeRenderer::drawClippedPolygon(vislib::math::Cuboid<float
 
     // check for each clip plane
     float vcpd;
-    for (int i = 0; i < this->volClipPlane.Count(); ++i) {
+    for (int i = 0; i < static_cast<int>(this->volClipPlane.Count()); ++i) {
         slices.setupSingleSlice(this->volClipPlane[i].PeekComponents(), position.PeekComponents());
         float d = 0.0f;
         vcpd = static_cast<float>(this->volClipPlane[i].PeekComponents()[3]);
